requirements.txt
lxml
requests
bs4
pandas
python-dateutil
python-telegram-bot 
pytesseract 
Pillow
opencv-python-headless
python-dotenv
openai
instaloader
fuzzywuzzy
python-Levenshtein